changed vector from int to double

+++ include/minimal_leastsq.h
@@ -48,6 +48,6 @@ struct LeastSq {
 };
 
 // Find the coefficient for the high-order term in the running time, by minimizing the sum of squares of relative error.
-LeastSq minimalLeastSq(const std::vector<int>& N, const std::vector<int>& Time, const BigO Complexity = BigO::O_Auto);
+LeastSq minimalLeastSq(const std::vector<int>& N, const std::vector<double>& Time, const BigO Complexity = BigO::O_Auto);
 
 #endif
